+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
+## 0.8.14
+
+- Filter out `only` suffix tag for getSwaggerFiles function
+- Throw error when readme file not found
+
 ## 0.8.13
+
 - Add rule INVALID_TYPESPEC_LOCATION to validate if TypeSpec file in 'resource-manager' or 'data-plane' folder.
 
 ## 0.8.12